% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_range.R
\name{cat_endash}
\alias{cat_endash}
\title{Catenate two characters by endash}
\usage{
cat_endash(x, y)
}
\arguments{
\item{x}{First arg in character}

\item{y}{Second arg in character}
}
\description{
Catenate two characters by endash
}
