% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_range.R
\name{print_decimal}
\alias{print_decimal}
\title{Print numeric in float with given decimal points}
\usage{
print_decimal(x, decimals)
}
\arguments{
\item{x}{Numeric vector to convert}

\item{decimals}{Number of decimal points to print}
}
\description{
Print numeric in float with given decimal points
}
