% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_range.R
\name{range4table}
\alias{range4table}
\title{Print range of given arg for table output}
\usage{
range4table(x, decimals = 1)
}
\arguments{
\item{x}{Numeric to calculate range}

\item{decimals}{Number of decimal points to print}
}
\description{
Print range of given arg for table output
}
